chess-ai-large-data/requirements.txt
torch>=2.2.0
python-chess>=1.999
zstandard>=0.22.0
pyarrow>=15.0.0
tqdm>=4.66.0
PyYAML>=6.0.1
requests>=2.31.0
numpy>=1.26.0
Pillow>=10.2.0
pygame>=2.6.0
